💩✨🐛: store correct LoadBalancer IP in service
This determines the correct external LoadBalancer IP from the discovered
prefixes, doing some black magic IP calculation. The code is prepared to
allocate new addresses and even prefixes, with lots of comments which
code is MVP-only and has to be nuked later (like the black magic IP
calculation).

Also checks if another Service in the cluster already uses the one
usable LoadBalancer IP and returns an error, which is added as event to
the Service, making it visible for the customer.