🐛: fix delete logic in LBaaS reconciliation
Previously, we looped through all steps until the first picked something
to create, we then created resources and restarted the loop. This made
problems with renaming resources, as that results in some resources to
be deleted and the same number of resources to be created. With later
steps creating resources depending on resources created in earlier
steps, the delete part never succeeded.

This commit changes the logic to "collect all to create and destroy",
fixing this problem and maybe even speeding things up a little.