This branch only disables github pages jekyll
Github pages previously used jekyll to render the pages and deployed
them itself. Now there is a action for Github Actions, which we use
instead. This branch only exists to disable old Github Pages handling,
both by disabling Jekyll and by having a branch that is never touched.
The branch is also not allowed to update the github-pages environment.