Rebuild fetching instance ID by name
We do not try to find the customer ID prefix anymore, instead using it
if configured or searching with a wildcard prefix if not. If no VM is
found with either prefix version, we try to find a VM with the exact
name (no prefix at all).

If at any point we find multiple VMs, we try to figure out the correct
one by looking at the IP addresses, with the InternalIP already being
set from the running node itself. If we still have multiple VMs at this
point, we give up.

This fixes ANXKUBE-640.