Add support for the PROXY protocol
By setting the hostname via an annotation on the given Service, we can
workaround the shenanigans of kube-proxy to enable support for the PROXY
protocol. In addition to the annotation being set, there's a
configuration change needed by the LBaaS operators which is done
manually right now.

This change is also done by other CCM implementations, like the one of
[DigitalOcean].

[DigitalOcean]: https://github.com/digitalocean/digitalocean-cloud-controller-manager/blob/d274be6e44f650311e4b960960b23001e6e513b4/cloud-controller-manager/do/loadbalancers.go#L236

Right now, this implementation is lacking proper tests to test this
behaviour, which were skipped due to missing test infrastructure and a
lack of time.

Closes: ANXKUBE-1251
Co-authored-by: Zofia Hagenguth <zhagenguth@anexia.com>