Expose the IP together with the hostname
Because the IP of the LoadBalancer Service is no longer exposed, it would
get harder to setup DNS records. The current workaround would be:

 * Wait for the LoadBalancer to be reconciled without the annotation.
 * Write down the IP address to configure it elsewhere.
 * Annotate the service.

This only works because there's a limitation of only one LoadBalancer
per cluster at the moment. Yet, it can turn out to get problematic for
any kind of automation.

As such, we're exposing the hostname together with the IP. This also
removes a lot of the code introduced in #355 and moves the code to its
previous implementation.

Closes: ANXKUBE-1280
See also: ANXKUBE-1251, #355